+++ yazi-core/src/cmp/option.rs
@@ -1,3 +1,4 @@
+use yazi_macro::impl_data_any;
 use yazi_shared::{Id, path::PathBufDyn, url::UrlBuf};
 
 use crate::cmp::CmpItem;
@@ -9,3 +10,5 @@ pub struct CmpOpt {
 	pub word:       PathBufDyn,
 	pub ticket:     Id,
 }
+
+impl_data_any!(CmpOpt);