+++ yazi-core/src/mgr/displace.rs
@@ -1,3 +1,4 @@
+use yazi_macro::impl_data_any;
 use yazi_shared::url::UrlBuf;
 
 #[derive(Clone, Debug)]
@@ -5,3 +6,5 @@ pub struct DisplaceOpt {
 	pub to:   Result<UrlBuf, yazi_fs::error::Error>,
 	pub from: UrlBuf,
 }
+
+impl_data_any!(DisplaceOpt);